+++ ignite/cmd/account_list.go
@@ -22,6 +22,7 @@ func accountListHandler(cmd *cobra.Command, _ []string) error {
 	ca, err := cosmosaccount.New(
 		cosmosaccount.WithKeyringBackend(getKeyringBackend(cmd)),
 		cosmosaccount.WithHome(getKeyringDir(cmd)),
+		cosmosaccount.WithBech32Prefix(getAddressPrefix(cmd)),
 	)
 	if err != nil {
 		return err